chore: move state management to useStore
